+++ src/posts/posts.service.ts
@@ -1,11 +1,22 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Inject,
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+  forwardRef,
+} from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PostsRepository } from './posts.repository';
+import { PublicationsService } from '../publications/publications.service';
 
 @Injectable()
 export class PostsService {
-  constructor(private readonly postsRepository: PostsRepository) {}
+  constructor(
+    private readonly postsRepository: PostsRepository,
+    @Inject(forwardRef(() => PublicationsService))
+    private readonly publicationsService: PublicationsService,
+  ) {}
 
   async create(createPostDto: CreatePostDto) {
     return await this.postsRepository.createPost(createPostDto);
@@ -28,6 +39,8 @@ export class PostsService {
 
   async remove(id: number) {
     await this.findOne(id);
+    const publication = await this.publicationsService.findOneByPostId(id);
+    if (publication) throw new UnauthorizedException();
     await this.postsRepository.deletePostById(id);
     return;
   }
